backend/college-directory/requirements.txt
streamlit
pandas
numpy
faiss-cpu
sentence-transformers
torch
scikit-learn
tqdm